Remove app themes
Since this is a library project and it doesn't have any activities
declared in the manifest file, these themes are useless. Also, this might
override the themes declared in the main project (that uses this library)
and, for example, break appcompat library usage. The app might crash with
IllegalStateException, telling the user that his/her activity should have
Theme.AppCompat theme set (which is exactly what happened in my case)